+++ src/WebJobs.Extensions.OpenAI/Search/ISearchProvider.cs
@@ -10,7 +10,10 @@ namespace Microsoft.Azure.WebJobs.Extensions.OpenAI.Search;
 /// </summary>
 public interface ISearchProvider
 {
-    string Name { get; set; }
+    /// <summary>
+    /// Name of the Search Provider.
+    /// </summary>
+    string Name { get; }
 
     /// <summary>
     /// Adds a document to a search provider index.