+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added dotnet-isolated support for Embeddings
-- Added AI Search support for Semantic Search.
+- Added Azure AI Search support for Semantic Search.
 
 ### Breaking changes